Add initial Savanna CLI
This commit provides the framework for the Savanna CLI. It is based on
the python-novaclient. novaclient was chosen because of its maturity
and the ubiquity in which it has been previously copied, i.e. we're in
the same boat as most everyone else. The goal is to keep modifications
in Savanna's copy to a minimum. To that end, much of the novaclient
code is retained, though commented out when it is not applicable.

A simple plugins-list/show command is provided with the framework for
demonstration and testing purposes.

Change-Id: I6bbbc86135d6cb16e088dd76261fdef29c2cb929
Implements: blueprint python-savannaclient-cli