JobExecutionsManager.create() should handle input_id/output_id == None
Input and output ids are not part of the Java workflow JSON schema. Including
them, even set to None, causes an error.  In order to maintain backward
compatibility, leave input_id and output_id as required parameters but
do not add them to the job exec data if they are None.

Closes-Bug: #1268695
Change-Id: I8df8cb86094f5d67ca16de7d6dd253af142eca25