Make the single entrypoint for client
It's a common approach to have such common Client with the following
path: "someclient.client.Client".

It contains mapping from client/API version to the class that handles it.

This change will make us able to perform any refactoring inside the client
w/o need to change paths in all projects that are using Savanna client.

Change-Id: I7d0237591ed7beccbd43fe60ce3b6cfc522fcf31