Fix duplicate help strings
Currently:
$ sahara help bash-completion
usage: sahara bash-completion

Prints all of the commands to stdout to support bash completion. Prints
all of the commands and options to stdout so that the sahara.bash_completion
script doesn't have to hard code them.

$ sahara help job-binary-data-create
usage: sahara job-binary-data-create [--file FILE]

Store data in the internal DB. Store data in the internal DB. Use 'swift
upload' instead of this command. Use this command only if Swift is not
available.

Update the strings to avoid duplication.

Change-Id: I4bb692026e1bfc2c210a976f3dfedb550d544f26