Removed cluster validation on client side
There is no sense in validation on client side. I believe we
don't want to have all validation logic in two places.

Also, the current validation was wrong for several reasons:
1. Wrong APIException usage (see bug description)
2. It required default_image_id for the cluster. Image could be
   set in each node group individually.
3. It required cluster_configs that is not mandatory

So, the only correct check was for node_groups field, but still
I don't think that client is the right place for checking such
things. If one day we will allow clusters with no nodegroups
client should allow that without modifications.

Change-Id: I3a7bf39b102b699e8f4a1e493314412cfd26ee2e
Closes-Bug: #1420569