API to mount manila shares to clusters
Manually mounting shares to every node in a large cluster would be a tedious
and error-prone process. Auto-mounting shares that are requested for use in
either the data source or binary storage case might be feasible for some use
cases. However, outside of our (optional) EDP interface this functionality
would never be usable. As such, it is best to provide the user an API for
mounting of shares onto Sahara clusters.

Change-Id: Ib897d3fefaea146150963e25fbaaec80e197ca36
Implements: blueprint mount-share-api