Designate integration
added 'domain_name' parameter to Cluster Template in API
and CLI client.

NOTE: in spec we proposed to use two domains: one for internal
resolution and another for external. But we decided to use only one
domain for both sides so an instance will have one hostname in the
domain and two appropriate A records because it's more convenient.

partial-bp: designate-integration
Change-Id: Ic636d1d9d27f9f96fbfd37b7d29d8f0add44a24f