Remove outdated tools and files
Those tools have been mostly removed from the repositories and
they do not add much to what can be done with few commands
(direct call of virtualenv, or use of tox for various test runners).
MANIFEST.in is not needed anymore apparently too, thanks
to changes in pbr (thanks Jeremy Freudberg for noticing this).

Change-Id: I4cfb47a758f39f2de5ebd77fc88124335ded4aa3