Fix the cover tox target (switch to stestr)
The cover tox target was broken for a while.
Switch it to use coverage+stestr.

Patch proposed again with the wrong authorship information.
For some reason I messed up them when I first sent out
the patch, and I did not notice.
If the author would have been the one reported in
the original patch, I would have simply merged it with
https://review.openstack.org/#/c/571480/

Change-Id: Ie907ec3b381fcc98955c6be4280caf3b5cb220d7